+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Upcoming Release
+
+- Add support for read mapping using `minimap2`.
+
 ## v0.4.1
 
 - Fix tree inference pipeline failed to capture `-T AUTO` option in `iqtree2`.